Fixed logs key

+++ migrations/2023-03-08-115547_starting_log/up.sql
@@ -1,3 +1,3 @@
 -- Your SQL goes here
 
-CREATE TABLE starting_logs (id int8 PRIMARY KEY, timestamp int8)
+CREATE TABLE starting_logs (id bigserial PRIMARY KEY, timestamp int8)
