+++ migrations/2023-03-08-115609_finished_log/up.sql
@@ -1,3 +1,3 @@
 -- Your SQL goes here
 
-CREATE TABLE finished_logs (id int8 PRIMARY KEY, timestamp int8)
+CREATE TABLE finished_logs (id bigserial PRIMARY KEY, timestamp int8)
